edits on irsensor, linefollower, HTHING, and pid. started qrd_encoder class

+++ lib/line_follower/linefollower.cpp
@@ -14,6 +14,7 @@ linefollower::linefollower(track *_left_track, track *_right_track, irsensor *_m
 
 void linefollower::follow_line()
 {
+	// Serial.println("8Start line following......");
 	//get and calculate the control error
 	main_sensor->update();
 	//float error = main_sensor->max_position() - 4.5;
@@ -27,8 +28,9 @@ void linefollower::follow_line()
 	{
 		//control the tracks normally (on solid ground)
 		
-		left_track->run(left_speed);
-		right_track->run(right_speed);
+		// Serial.println("No CLIFF!!!");
+		left_track->run((int)left_speed);
+		right_track->run((int)right_speed);
 	}
 	else {
 		//driving off a cliff, continue to go forward
